+++ src/main/java/com/mysql/MySqlTicketDao.java
@@ -16,6 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManagerFactory;
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
@@ -108,7 +109,7 @@ public class MySqlTicketDao implements TicketDao {
     }
 
     @Override
-    public Ticket getTicketById(int id) throws DaoException {
+    public Ticket getTicketById(BigInteger id) throws DaoException {
         try {
             session = sessionFactory.openSession();
             session.beginTransaction();
