+++ src/main/java/com/web/TicketEditor.java
@@ -4,6 +4,7 @@ import com.domain.Ticket;
 import org.springframework.stereotype.Component;
 
 import java.beans.PropertyEditorSupport;
+import java.math.BigInteger;
 
 @Component
 public class TicketEditor extends PropertyEditorSupport {
@@ -12,7 +13,7 @@ public class TicketEditor extends PropertyEditorSupport {
     public void setAsText(String text) {
         try {
             Ticket t = new Ticket();
-            t.setTicketId(Integer.parseInt(text));
+            t.setTicketId(BigInteger.valueOf(Integer.parseInt(text)));
             this.setValue(t);
         } catch (Exception e) {
             e.printStackTrace();
